@@ -6,8 +6,8 @@ build-backend = "setuptools.build_meta"
 [project]
 dynamic = ["dependencies", "optional-dependencies"]
 name = 'boxes'
-version = '1.0'
-requires-python = '>=3.8'
+version = '1.1'
+requires-python = '>=3.9'
 description = 'Boxes generator for laser cutters'
 maintainers = [ { name = 'Florian Festi', email='florian@festi.info' } ]
 readme = "README.rst"
