@@ -13,7 +13,6 @@
 #   You should have received a copy of the GNU General Public License
 #   along with this program.  If not, see <http://www.gnu.org/licenses/>.
 from collections import OrderedDict
-from typing import Tuple
 
 from boxes import *
 
@@ -133,7 +132,7 @@ In Use:
 
     def _calc_grid_size_width_offset(
         self, inner_mm_after_nesting: float
-    ) -> Tuple[int, float]:
+    ) -> tuple[int, float]:
         """Return the size of the grid and the offset from the outer top right corner"""
         hole_count = self._calc_hole_count(inner_mm_after_nesting)
         grid_size = (
